# Refactor of FastHistMaker to allow for custom regularisation methods