--- a/src/tree/param.h
+++ b/src/tree/param.h
@@ -12,6 +12,7 @@
 #include <cmath>
 #include <cstring>
 #include <limits>
+#include <string>
 #include <vector>