--- a/src/tree/split_evaluator.cc
+++ b/src/tree/split_evaluator.cc
@@ -1,0 +1,156 @@
+/*!
+ * Copyright 2018 by Contributors
+ * \file split_evaluator.cc
+ * \brief Contains implementations of different split evaluators.
+ */
+#include <dmlc/registry.h>
+#include "split_evaluator.h"
+#include "param.h"
+#include "../common/host_device_vector.h"
+
+namespace dmlc {
+DMLC_REGISTRY_ENABLE(::xgboost::tree::SplitEvaluatorReg);
+} // namespace dmlc
+
+namespace xgboost {
+namespace tree {
+
+SplitEvaluator* SplitEvaluator::Create(const std::string& name) {
+  auto* e = ::dmlc::Registry< ::xgboost::tree::SplitEvaluatorReg>::Get()->Find(name);
+  if (e == nullptr) {
+    LOG(FATAL) << "Unknown SplitEvaluator " << name;
+  }
+  return (e->body)();
+}
+
+// Default implementations of some virtual methods that that need to do anything by default
+SplitEvaluator::~SplitEvaluator() {}
+void SplitEvaluator::Init(const std::vector<std::pair<std::string, std::string> >& args) {}
+void SplitEvaluator::Reset() {}
+
+bst_float SplitEvaluator::ComputeSplitLoss(bst_uint parentID,
+                                           bst_uint featureID,
+                                           const GradStats& left,
+                                           const GradStats& right) const {
+  return ComputeLoss(parentID, left) + ComputeLoss(parentID, right);
+}
+
+/*! \brief Encapsulates the parameters required by the RidgePenalty split evaluator */
+struct RidgePenaltyParams : public dmlc::Parameter<RidgePenaltyParams> {
+  float reg_lambda;
+  float reg_gamma;
+
+  DMLC_DECLARE_PARAMETER(RidgePenaltyParams) {
+    DMLC_DECLARE_FIELD(reg_lambda)
+      .set_lower_bound(0.0)
+      .set_default(1.0)
+      .describe("L2 regularization on leaf weight");
+    DMLC_DECLARE_FIELD(reg_gamma)
+      .set_lower_bound(0.0f)
+      .set_default(0.0f)
+      .describe("Cost incurred by adding a new leaf node to the tree");
+    DMLC_DECLARE_ALIAS(reg_lambda, lambda);
+    DMLC_DECLARE_ALIAS(reg_gamma, gamma);
+  }
+};
+
+DMLC_REGISTER_PARAMETER(RidgePenaltyParams);
+
+/*! \brief Applies an L2 penalty and per-leaf penalty. */
+class RidgePenalty : public SplitEvaluator {
+ public:
+  void Init(const std::vector<std::pair<std::string, std::string> >& args) override {
+    m_params.InitAllowUnknown(args);
+  }
+
+  SplitEvaluator* GetHostClone() const override {
+    RidgePenalty* r = new RidgePenalty();
+    r->m_params = this->m_params;
+
+    return r;
+  }
+
+  bst_float ComputeLoss(bst_uint parentID, const GradStats& stats) const override {
+    return (stats.sum_grad * stats.sum_grad) / (stats.sum_hess + m_params.reg_lambda) - m_params.reg_gamma;
+  }
+
+  bst_float ComputeWeight(bst_uint parentID, const GradStats& stats) const override {
+    return -stats.sum_grad / (stats.sum_hess + m_params.reg_lambda);
+  }
+
+ private:
+  RidgePenaltyParams m_params;
+};
+
+XGBOOST_REGISTER_SPLIT_EVALUATOR(RidgePenalty, "ridge")
+.describe("Use an L2 penalty term for the weights and a cost per leaf node")
+.set_body([]() {
+    return new RidgePenalty();
+  });
+
+  /*! \brief Encapsulates the parameters required by the MonotonicConstraint split evaluator */
+struct MonotonicConstraintParams : public dmlc::Parameter<MonotonicConstraintParams> {
+  std::vector<int> monotone_constraints;
+
+  DMLC_DECLARE_PARAMETER(MonotonicConstraintParams) {
+    DMLC_DECLARE_FIELD(monotone_constraints)
+      .set_default(std::vector<int>())
+      .describe("Constraint of variable monotonicity");
+    DMLC_DECLARE_ALIAS(reg_lambda, lambda);
+    DMLC_DECLARE_ALIAS(reg_gamma, gamma);
+  }
+};
+
+/*! \brief Uses the same criteria as RidgePenalty, with the addition that the resulting tree must be monotonically
+      increasing/decreasing with respect to a user specified set of features.
+*/
+class MonotonicConstraint : public RidgePenalty {
+ public:
+  void Init(const std::vector<std::pair<std::string, std::string> >& args) override {
+    m_params.InitAllowUnknown(args);
+
+    RidgePenalty::Init(args);
+  }
+
+  SplitEvaluator* GetHostClone() const override {
+    MonotonicConstraint* c = new MonotonicConstraint();
+    c->m_params = this->m_params;
+
+    return c;
+  }
+
+  bst_float ComputeSplitLoss(bst_uint parentID,
+                             bst_uint featureID,
+                             const GradStats& left,
+                             const GradStats& right) const override {
+
+    bst_float infinity = std::numeric_limits<bst_float>::infinity();
+    bst_uint constraint = m_params.monotone_constraints[featureID];
+
+    bst_float loss = ComputeLoss(parentID, left) + ComputeLoss(parentID, right);
+    bst_float leftWeight = ComputeWeight(parentID, left);
+    bst_float rightWeight = ComputeWeight(parentID, right);
+
+    if(constraint == 0) {
+      return loss;
+    }
+    else if(constraint > 0) {
+      return leftWeight <= rightWeight ? loss : infinity;
+    }
+    else {
+      return leftWeight >= rightWeight ? loss : infinity;
+    }
+  }
+
+ private:
+  MonotonicConstraintParams m_params;
+};
+
+XGBOOST_REGISTER_SPLIT_EVALUATOR(MonotonicConstraint, "monotonic_constraint")
+.describe("Enforces that the tree is monotonically increasing/decreasing w.r.t. specified features")
+.set_body([]() {
+    return new MonotonicConstraint();
+  });
+
+} // tree
+} // xgboost