… CONTRIBUTORS.md
… src/tree/split_evaluator.h
… src/tree/updater_colmaker.cc
… src/tree/updater_fast_hist.cc