Currently only FastHistMaker has been modified to use this new regularisation framework. I figured I should get some feedback before I put in the effort of refactoring the other updaters.

Future PRs will focus on novel regularisation methods.

**Things that need to be checked:**
- [x] L2 weight penalty is applied correctly
- [x] Monotonic constraints are applied correctly
- [x] Speed is not impacted noticeably