@hcho3 2018-05-23
Nice! I like how split evaluation logic is now being separated out from the updaters. There is a pending PR (#3136) to implement feature interaction constraints. Do you see a way to incorporate feature interaction constraints in the SplitEvaluator construct?