@henrygouk 2018-05-25
@hcho3 It should be possible to incorporate that into this framework. The efficient way to do it would be to add another method to the SplitEvaluator that is called each time a split is applied to the tree, which would update some internal state of the SplitEvaluator. E.g., which features have already been split on.