@BlueTea88 2018-05-25
I might be wrong, but it looks like the monotonicity constraints only holds at the local split level.  It might not guarantee monotonicity for subsequent splits of the same tree which the current code does - through lower bounds and upper bounds?