@hcho3 2018-06-03
@henrygouk Are you dropping `AdditiveSplitEvaluator`? If so, we may be able to eliminate virtual function calls by having `SplitEvaluator` as a template argument to the tree updater. See `TConstraint` for an existing example. (In particular, `spliteval_` should be of type `SplitEvaluator`, not `std::unique_ptr<SplitEvaluator>.) Only downside is that the updater code gets duplicated by however many evaluators there are. If you'd like, I can propose a template version of your code and you can run some benchmark.