@henrygouk 2018-06-05
If the code currently in master is using the templated approach would it be suitable to just compare the performance of this PR to master?