@RAMitchell 2018-06-06
@hcho3 I prefer virtual functions unless there is a visible speed difference.