@henrygouk 2018-06-06
using the benchmarking script, tests/benchmark/benchmark_tree.py, I've come up with these numbers

This PR:
tree_method=hist Train Time: 60.4643552303 seconds
tree_method=exact Train Time: 586.678107023 seconds

master:
tree_method=hist Train Time: 66.2623929977 seconds
tree_method=exact Train Time: 621.636813879 seconds