@RAMitchell 2018-06-10
There are some failures on Windows relating to the dumped tree outputs: 
`` == testthat results  ===========================================================
  OK: 532 SKIPPED: 4 FAILED: 7
  1. Failure: xgb.dump works (@test_helpers.R#44) 
  2. Failure: xgb.dump works (@test_helpers.R#53) 
  3. Failure: xgb.model.dt.tree works with and without feature names (@test_helpers.R#213) 
  4. Failure: xgb.importance works with and without feature names (@test_helpers.R#239) 
  5. Failure: updating the model works (@test_update.R#32) 
  6. Failure: updating the model works (@test_update.R#41) 
  7. Failure: updating the model works (@test_update.R#64)```

You can find these tests here: https://github.com/dmlc/xgboost/tree/master/R-package/tests/testthat
Curiously these tests pass on Linux. 

If this can be solved I think this PR is mergable.