@hcho3 2018-06-22
@henrygouk Curiously, the R-MinGW test fails for Windows R 32-bit. Investigating why.