@hcho3 2018-06-22
@henrygouk @RAMitchell It turns out that the split evaluator produces a **substantially different** model when XGBoost is compiled for 32-bit instead of the usual 64-bit. I'm attaching the model dumps made on 32-bit and 64-bit.
[win32.txt](https://github.com/dmlc/xgboost/files/2129390/win32.txt)
[win64.txt](https://github.com/dmlc/xgboost/files/2129391/win64.txt)