@henrygouk 2018-06-27
I think I've figured out the problem.

By changing all occurrences of 'bst_float' in split_evaluator.cc and split_evaluator.h to be 'double' I can get the R tests to pass when running in 32-bit mode. This is because the x87 floating point unit is used in the x86 ISA, whereas in AMD64 the scalar SSE instructions are used. Something that is kind of annoying is that all floating point arithmetic done with the x87 unit uses 80 bit registers. The values are then rounded when stored in the 32/64 float/double variables in memory. Therefore, expressions will carried out with much higher precision in 32-bit mode, and the precision loss will only occur when the values are stored in memory.

Changing the bst_float occurrences to double is not really a good solution to this problem. 64-bit floating point arithmetic will certainly be closer to the 80-bit arithmetic done using the x87 units, but it is still possible to get different answers. I think the correct way to solve this problem would be to find a way to build the 32-bit R library with SSE2 enabled, however I have absolutely no experience with building R bindings, so perhaps someone else should look into this.