@hcho3 2018-06-27
> I think the correct way to solve this problem would be to find a way to build the 32-bit R library with SSE2 enabled.

I've added `-msse2` flag to Win32 R build. Hope it solves the problem.