@hcho3 2018-06-27
@hetong007 I am worried that adding `-msse2` for Win32 would break installation for users of old machines that do not support SSE2. Are we committed to supporting machines without SSE2?