@hetong007 2018-06-27
I'm not award of SSE2 related regulation from CRAN. Maybe it is good to test it through https://win-builder.r-project.org/.

The best practice could be that we by default compile with SSE2, and provide a switch in `Makevars.win` for users who want to compile without it.