@hcho3 2018-06-27
@hetong007 It appears that most Windows machines today would have SSE2-capable CPUs (see [this article](https://www.gizmodo.com.au/2018/06/microsoft-quietly-drops-support-for-non-sse2-cpus-in-windows-7/) about Windows 7 dropping support for non-SSE2 machines).

How does a typical R user specify a switch for compilation? Does `install.packages` or `devtools::install()` let the user pass extra flags?