@hcho3 2018-06-28
@henrygouk I rebased the whole branch against the latest master, so that the GPU tests would pass. Once all tests pass, I will merge the PR. Thanks for your contribution.