@hetong007 2018-06-28
@hcho3 that sounds good still. I think we can always take a look back if that raise a significant issue.