@hcho3 2018-06-28
@BlueTea88 @henrygouk Now that this PR has been merged, let's discuss how we can add feature interaction constraints.