@BlueTea88 2018-06-28
@hcho3 Awesome, I can try drafting up a new method for interaction constraints in split_evaluator as @henrygouk mentioned.